Fix the "notes missing when repeating"-bug.
Found out what that bug is about, even though I do not
really understand it.  It appears that, while there is a
"lighting" note on and another midi note on comes from keyboard,
if the note off event occurs (user pulls a finger away from key),
another note on event will happen at that moment with velocity 0.
This note on event would again be counted as something that should
move the playback forward, which is not supposed to happen,
causing the effect of "missing notes" when those notes are
something that is expected.

This behaviour may be specific to Yamaha EZ-220, unfortunately.
It would be nice to be able to test with other midi keyboards
with similar features.

Anyway, it appears the bug is finally fixed.