Trying to do something to fix playback issues.
Yep... use non-blocking device descriptor for midi device.
Initialize structures for poll() each time it is called,
with mio_pollfd(), and use mio_revents() to look up
if there are pending events.  Do not call poll() with timeout 0,
instead just skip to playback, calling poll() should be useless
in this case.

This does not fix the issues I am seeing with background playback,
where all notes appear to be squashed to be played immediately
instead of in time.  Nevertheless, these changes to do not look
wrong to me, so I will go with these...