Compare timing between lighted keys and playback.
Check out the ticks of lighted key events and playback midi
events, and do not progress if playback would progress beyond
the event timings of of lighted keys.

Without this there would be a problem as midi events are not
sorted such that playback events would be always after lighted
key events.  Sorting the events in a different way could solve
this in a different way, but this is probably cleaner.

There are still issues with playback timing, though.
The mystery deepens...